requirements.txt
streamlit==1.32.0
opencv-python-headless==4.9.0.80
facenet-pytorch==2.5.3
torch==2.2.0
torchvision==0.17.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
pillow==10.2.0
